+++ src/olympia/landfill/generators.py
@@ -6,12 +6,12 @@ from itertools import cycle, islice
 
 from django.db.models.signals import post_save
 
-from olympia.addons.forms import icons
 from olympia.addons.models import Addon, Persona, update_search_index
 from olympia.amo.utils import slugify
 from olympia.constants.applications import APPS, FIREFOX
 from olympia.constants.base import (
     ADDON_EXTENSION, ADDON_PERSONA, ADDON_STATICTHEME, STATUS_PUBLIC)
+from olympia.devhub.forms import icons
 
 from .categories import generate_categories
 from .collection import generate_collection
